refactored into a common BaseSampler with two implementation for RF and BDT
